% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_time_periods.R
\name{determine_time_periods}
\alias{determine_time_periods}
\title{Determine Time Periods}
\usage{
determine_time_periods(start_date, end_date, num_of_days, time_window)
}
\arguments{
\item{start_date}{The starting date of the time period.}

\item{end_date}{The ending date of the time period.}

\item{num_of_days}{Number of days between \code{start_date} and \code{end_date}.}

\item{time_window}{Time window in days to divide the period into segments.}
}
\value{
A vector of dates representing the determined time periods.
}
\description{
Helper function to calculate time periods based on the given start date, end date,
number of days, and time window.
}
