% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_request_trends.R
\name{initialize_request_trends}
\alias{initialize_request_trends}
\title{Initialize Google Trends Request}
\usage{
initialize_request_trends(
  keyword,
  topic = NULL,
  folder_name,
  start_date,
  end_date,
  data_format
)
}
\arguments{
\item{keyword}{The keyword to be used for collecting Google Trends data.}

\item{topic}{The topic associated with the keyword. For example, '/m/0ddwt' will give Google Trends data for Insomnia as topic of 'Disorder'. If identical to the keyword, data will reflect Google Trends search term data. NOTE: URL's have certain codes for special characters. For example, \%20 = white space, \%2F = / (forward slash) etc.}

\item{folder_name}{Name of the parent folder where all data will be stored.}

\item{start_date}{The start date from which to collect Google Trends data.}

\item{end_date}{The end date until which to collect Google Trends data.}

\item{data_format}{Time basis for the query. Can be one of 'daily', 'weekly', or 'monthly'.}
}
\value{
A list containing initialized values and objects for further interaction with the package:
\item{logger}{A logging object for recording messages.}
\item{keyword}{The keyword used for data collection.}
\item{topic}{The topic associated with the keyword.}
\item{folder_name}{Name of the parent folder for storing data.}
\item{start_date}{Start date for data collection.}
\item{end_date}{End date for data collection.}
\item{data_format}{Time basis for the data query ('daily', 'weekly', or 'monthly').}
\item{num_of_days}{Number of days between \code{start_date} and \code{end_date}.}
\item{pytrend}{Initialized pytrends request object.}
\item{time_window}{Optional. Time window parameter, applicable for 'weekly' data format.}
\item{times}{Optional. Time periods determined for 'weekly' or 'daily' data formats.}
}
\description{
This function initializes a request to the Google Trends API using pytrends,
creates necessary directories, and prepares parameters for data collection.
}
\details{
The initiation stage involves creating two folders automatically:
- The main folder chosen by the user (`folder_name`).
- A subfolder corresponding to the `data_format` (e.g., 'daily', 'weekly', 'monthly') for storing data.
}
\examples{
# Create a temporary folder for the example

# Ensure the temporary folder is cleaned up after the example
if (reticulate::py_module_available("pytrends")) {
  # Run the function with the temporary folder
  params <- initialize_request_trends(
    keyword = "Coronavirus disease 2019",
    topic = "/g/11j2cc_qll",
    folder_name = file.path(tempdir(), "test_folder"),
    start_date = "2024-05-01",
    end_date = "2024-05-03",
    data_format = "daily"
  )
  on.exit(unlink("test_folder", recursive = TRUE))
} else {
  message("The 'pytrends' module is not available.
  Please install it by running install_pytrendslongitudinalr()")
}

}
