% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pytrendslongitudinalr}
\alias{install_pytrendslongitudinalr}
\title{Install and Set Up Python Environment for PytrendsLongitudinalR}
\usage{
install_pytrendslongitudinalr(envname = "pytrends-in-r-new", ...)
}
\arguments{
\item{envname}{Name of the virtual environment.}

\item{...}{Additional arguments passed to `py_install()`.}
}
\description{
This function sets up the Python virtual environment and installs required packages.
}
