% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_nmonthly.R
\name{time_series_nmonthly}
\alias{time_series_nmonthly}
\title{Time Series Data Collection Method for 'weekly/daily'}
\usage{
time_series_nmonthly(params, reference_geo_code = "US")
}
\arguments{
\item{params}{A list containing parameters including logger, pytrend, folder_name, data_format, start_date, end_date, keyword, and topic.}

\item{reference_geo_code}{Geographic code to specify the location for data collection (default is "US").}
}
\description{
This function collects time series data on a weekly/daily basis using the Google Trends API.
It retrieves interest over time data using pytrends.interest_over_time() and saves it to the specified directory.
}
\details{
This function collects Google Trends time series data based on the specified parameters and saves it in the following structure:
\code{folder_name/data_format/over_time/reference_geo_code}.
The filename includes the start and end dates of the data period.
}
\keyword{internal}
