% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_time_series.R
\name{concat_time_series}
\alias{concat_time_series}
\title{Concatenation of Time series method}
\usage{
concat_time_series(params, reference_geo_code = "US", zero_replace = 0.1)
}
\arguments{
\item{params}{A list containing parameters including logger, folder_name, data_format, time_window, and other necessary parameters for sub-functions.}

\item{reference_geo_code}{This is the same geo code that is used in collecting time_series data. If the time_series data for that geo is not collected beforehand, or the file does not exist, it will throw and error. Default is 'US'}

\item{zero_replace}{As data from different time periods are rescaled, sometimes the last/first data point of a period might be zero. Then the calculation will throw error or everything single data point will become zero. To avoid that, we are tweaking the zeroes to be of an insignificant number to carry on with the calculation.}
}
\description{
Concatenation of Time series method
}
\details{
This method will concat the time series data collected in time_series() method.
Because the data points in time_series is independent of each other, they needs to be re-aligned to get correct index for the given time period.
This method concatenates time_series data for all the period and gives back the combined rescaled time_series data for the reference timeline.
This rescaled time_series data will be used in the next method to rescale the cross_section data.
}
\examples{
\dontrun{
# Set up a temporary directory
temp_dir <- tempdir()
folder_name <- file.path(temp_dir, "biden")

# Example usage
params <- initialize_request_trends("Joe Biden", "/m/012gx2",
folder_name, "2017-12-3", "2024-05-1", "weekly")
time_series(params, reference_geo_code = "US-CA")
concat_time_series(params, reference_geo_code = "US-CA")

# Clean up temporary directory
unlink(temp_dir, recursive = TRUE)
}
}
