% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_section.R
\name{cross_section}
\alias{cross_section}
\title{Collect Cross Section Data}
\usage{
cross_section(params, geo = "", resolution = "COUNTRY")
}
\arguments{
\item{params}{A list containing parameters including logger, data_format, folder_name, start_date, end_date, keyword, topic, num_of_days, and pytrend.}

\item{geo}{Country/Region to collect data from. Defaults to worldwide if empty.}

\item{resolution}{Resolution level for the data ('COUNTRY', 'REGION', 'CITY'). Defaults to 'COUNTRY'.}
}
\description{
This function collects cross section data for a given keyword and timeline using the Google Trends API.
It retrieves data using the pytrends.interest_by_region() method and saves it automatically to the specified directory.
}
\details{
This function collects Google Trends data based on the specified parameters and saves it in the following structure:
\code{folder_name/data_format/by_region}.
Each file contains data for a specific time period (day/week/month) and geographical region.
The filenames include the start and end dates of the data period.

PS: This method may take a long time to complete due to Google Trends API rate limits.
For example, collecting 350 days of daily data can take approximately 5 hours.
}
\examples{
\donttest{
# Please note that this example may take a few minutes to run
# Create a temporary folder for the example

# Ensure the temporary folder is cleaned up after the example
# Run the function with the temporary folder
params <- initialize_request_trends(
  keyword = "Joe Biden",
  topic = "/m/012gx2",
  folder_name = file.path(tempdir(), "biden_save"),
  start_date = "2024-05-01",
  end_date = "2024-05-03",
  data_format = "daily"
)

# Run the cross_section function with the parameters
tryCatch({
  cross_section(params, geo = "US", resolution = "REGION")
}, pytrends.exceptions.TooManyRequestsError = function(e) {
  message("Too many requests error: ", conditionMessage(e))
})
on.exit(unlink(tempdir(), recursive = TRUE))
}
}
